refactor: new finalize and llm modules
